LayoutTests/http/wpt/webauthn/public-key-credential-create-success-ccid.https-expected.txt

CONSOLE MESSAGE: User gesture is not detected. To use the WebAuthn API, call 'navigator.credentials.create' or 'navigator.credentials.get' within user activated events.

PASS PublicKeyCredential's [[create]] with minimum options in a mock ccid authenticator.
PASS PublicKeyCredential's [[create]] with minimum options in a mock ccid authenticator with contactless.
